Add files via upload
Added maya hammer file and the final exproted hammer. Didn't like the clash between the lower poly head of the hammer and the higher poly handle, so I beveled the edges of the head of the hammer to try and make it conflict less aesthetically. Still a little unfamiliar with combining/merging meshes into a single mesh, as such this model is composed of several pieces.